+++ src/main/java/com/sidematch/backend/config/oauth/OAuth2SuccessHandler.java
@@ -1,7 +1,7 @@
 package com.sidematch.backend.config.oauth;
 
 import com.sidematch.backend.config.jwt.TokenProvider;
-import com.sidematch.backend.config.oauth.oauth2user.CustomOAuth2User;
+import com.sidematch.backend.config.oauth.userservice.CustomOAuth2User;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
